routes-api/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
bcrypt==4.1.1
black==23.11.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.6.1
dnspython==2.4.2
email-validator==1.3.1
fastapi==0.95.0
frozenlist==1.4.0
graphql-core==3.2.3
greenlet==3.0.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.310
langsmith==0.0.43
libcst==1.1.0
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
motor==3.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.1
passlib==1.7.4
pathspec==0.11.2
pep8==1.7.1
platformdirs==4.0.0
pycodestyle==2.10.0
pydantic==1.10.7
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.5.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0
requests==2.31.0
rich==13.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.26.1
strawberry-graphql==0.209.6
tabulate==0.9.0
tenacity==8.2.3
tomli==2.0.1
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0
yarl==1.9.2
